backend/ml_service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pytesseract==0.3.10
imagehash==4.3.1

# Optional: For real ML models (uncomment when ready)
# transformers==4.35.0
# torch==2.1.0
# sentence-transformers==2.2.2
# clip-by-openai==1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

